refactor scraper into modules
