+++ backend/apps/imports/scrapers/viernulvier.py
@@ -1,46 +1,48 @@
-"""Viernulvier / Peppered scraper
+"""Compatibility facade for the Viernulvier scraper.
 
-Key design decisions
---------------------
-- One requests.Session per sync run (connection pooling, shared auth header).
-- Retry with exponential backoff + jitter for 429 / 5xx / network errors.
-- FK resolution via a warm in-memory cache (one bulk query per model) to
-  avoid the classic N+1 problem.
-- Concurrent page fetching via ThreadPoolExecutor (page order preserved).
-- Translation updates are *batched per language* - all translated fields for
-  one parent + one language are merged into a single update_or_create call
-  instead of one call per field (13 fields x 3 languages = 3 queries, not 39).
-  defined by multiple FK columns (e.g. EventPrice(event, price_rank)).
-- Per-item savepoints so one bad record never aborts the whole batch.
-- ETag / 304 support to skip completely unchanged endpoints.
-- Dry-run mode for safe inspection before writing.
-- Media item crop sync: fetches each foto MediaItem individually to retrieve
-  its embedded crops list, downloads the wanted variants (hd_ready, FE3_header)
-  and stores them locally via Django's ImageField / storage backend.
+The implementation is split over focused modules in this package:
+- `viernulvier_constants.py`
+- `viernulvier_http.py`
+- `viernulvier_normalize.py`
+- `viernulvier_relations.py`
+- `viernulvier_sync.py`
+- `viernulvier_media.py`
+
+This file intentionally keeps legacy symbol names and monkeypatch points
+used by the test suite and management command integrations.
 """
 
 from __future__ import annotations
 
 import logging
-import os
-import random
-import re
-import sys
 import time
-from collections import defaultdict
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from dataclasses import dataclass, field
-from decimal import Decimal, InvalidOperation
-from typing import Any, Callable, Dict, List, Mapping, Optional, Set, Tuple, Type
-from urllib.parse import urljoin, urlparse
+from typing import Any, Callable, Dict, List, Mapping, Optional, Tuple, Type
 
 import requests
-from django.core.exceptions import FieldDoesNotExist, FieldError, ValidationError
-from django.core.files.base import ContentFile
-from django.core.validators import URLValidator
-from django.db import DatabaseError, IntegrityError, models, transaction
+from django.db import models, transaction
 from django.utils import timezone
-from django.utils.dateparse import parse_date, parse_datetime
+from django.utils.dateparse import parse_datetime
+
+from . import viernulvier_constants as _constants
+from . import viernulvier_http as _http
+from . import viernulvier_media as _media
+from . import viernulvier_normalize as _normalize
+from . import viernulvier_relations as _relations
+from . import viernulvier_sync as _sync
+from .viernulvier_constants import M2MConfig, ModelSyncConfig, RateLimitError, ScraperError, TranslationConfig
+
+BASE_DOMAIN = _constants.BASE_DOMAIN
+BASE_URL = _constants.BASE_URL
+DEFAULT_ENDPOINT = _constants.DEFAULT_ENDPOINT
+DEFAULT_TIMEOUT = _constants.DEFAULT_TIMEOUT
+ERROR_CONTEXT_PATH = _constants.ERROR_CONTEXT_PATH
+MAX_CONCURRENT_PAGES = _constants.MAX_CONCURRENT_PAGES
+MAX_ERROR_MESSAGES = _constants.MAX_ERROR_MESSAGES
+MAX_RETRIES = _constants.MAX_RETRIES
+RETRY_BACKOFF_BASE = _constants.RETRY_BACKOFF_BASE
+RETRY_BACKOFF_MAX = _constants.RETRY_BACKOFF_MAX
+RETRY_STATUS_CODES = _constants.RETRY_STATUS_CODES
+USER_AGENT_POOL = _constants.USER_AGENT_POOL
 
 logger = logging.getLogger(__name__)
 
@@ -61,196 +63,30 @@ __all__ = [
     "clean_string",
 ]
 
-# ---------------------------------------------------------------------------
-# API configuration
-# ---------------------------------------------------------------------------
 
-BASE_URL = "https://www.viernulvier.gent/api/v1"
-BASE_DOMAIN = "https://www.viernulvier.gent"
-DEFAULT_ENDPOINT = "/productions"
-DEFAULT_TIMEOUT = 15
-ERROR_CONTEXT_PATH = "/api/contexts/Error"
+FKCache = _relations.FKCache
 
-# Retry settings
-MAX_RETRIES = 5
-RETRY_BACKOFF_BASE = 1.5  # seconds
-RETRY_BACKOFF_MAX = 60  # hard ceiling in seconds
-RETRY_STATUS_CODES = {429, 500, 502, 503, 504}
-
-# Concurrent page fetching
-MAX_CONCURRENT_PAGES = 4
-
-# Maximum error messages stored per sync run.
-# Prevents unbounded memory growth when thousands of records fail (e.g. 11k event prices).
-MAX_ERROR_MESSAGES = 50
-
-USER_AGENT_POOL = [
-    "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/122.0.0.0 Safari/537.36",
-    "Mozilla/5.0 (Macintosh; Intel Mac OS X 13_6_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/121.0.0.0 Safari/537.36",
-    "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:123.0) Gecko/20100101 Firefox/123.0",
-    "Mozilla/5.0 (Macintosh; Intel Mac OS X 14.3; rv:122.0) Gecko/20100101 Firefox/122.0",
-]
-
-# String values treated as empty *only* in URL fields.
-# Not used for regular CharField / TextField - "0" is valid text.
-_EMPTY_URL_VALUES: Set[str] = {"", "0", "none", "null", "undefined", "-", "n/a", "nvt"}
-
-
-# ---------------------------------------------------------------------------
-# Exceptions
-# ---------------------------------------------------------------------------
-
-
-class ScraperError(Exception):
-    """Raised when the scraper cannot fetch or normalize Viernulvier data."""
-
-
-class RateLimitError(ScraperError):
-    """Raised on HTTP 429 after all retries are exhausted."""
-
-    def __init__(self, retry_after: Optional[int] = None) -> None:
-        self.retry_after = retry_after
-        msg = f"Rate limited by API (Retry-After: {retry_after}s)" if retry_after else "Rate limited by API"
-        super().__init__(msg)
-
-
-# ---------------------------------------------------------------------------
-# Configuration dataclasses
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class TranslationConfig:
-    """Configuration for syncing one translated field to a translation model.
-
-    The Peppered API returns translations as flat dicts:
-        "title": {"nl": "De titel", "en": "The title"}
-
-    Each TranslationConfig handles one such field. Multiple configs that share
-    the same translation model are *merged per language* by _sync_all_translations
-    so only one update_or_create per language is needed - not one per field.
-
-    Args:
-        api_key:          Key in the parent API item (e.g. "title").
-        model:            Django translation model class.
-        parent_fk:        FK field name pointing to the parent on the translation model.
-        flat_field:       Field name on the translation model to populate.
-        language_fk:      Field name for the language code. Use "language_id" for
-                          models where Language has a string primary key.
-        value_transforms: Optional per-field callables applied after type coercion.
-    """
-
-    api_key: str
-    model: Type[models.Model]
-    parent_fk: str
-    flat_field: str
-    language_fk: str = "language_id"
-    value_transforms: Dict[str, Callable[[Any], Any]] = field(default_factory=dict)
-
-
-@dataclass
-class M2MConfig:
-    """Configuration for a M2M relation expressed via an explicit through table.
-
-    The API returns M2M data as a list of URL strings:
-        "genres": ["https://.../genres/1", "https://.../genres/3"]
-
-    Args:
-        api_key:              Key in the parent API item.
-        related_model:        The related Django model.
-        through_model:        The through/junction table model.
-        parent_fk:            FK to the parent model in the through table.
-        related_fk:           FK to the related model in the through table.
-        related_lookup_field: Field used to look up the related object (default: "external_id").
-        extra_fields:         Extra fields on the through table: {api_field: model_field}.
-                              Use "position" to auto-fill from the list index.
-    """
-
-    api_key: str
-    related_model: Type[models.Model]
-    through_model: Type[models.Model]
-    parent_fk: str
-    related_fk: str
-    related_lookup_field: str = "external_id"
-    extra_fields: Dict[str, str] = field(default_factory=dict)
-
-
-@dataclass
-class ModelSyncConfig:
-    """Complete sync configuration for one Django model.
-
-    Args:
-        field_map:        API key -> model field name. Set value to None to
-                          explicitly skip an API field.
-        value_transforms: Callables applied after type coercion, keyed by model
-                          field name. E.g. {"is_own_location": nee_ja_to_bool}.
-        fk_resolvers:     Custom FK resolvers when the default external_id lookup
-                          does not apply. Callable(raw_value) -> pk | None.
-        translations:     TranslationConfig list for flat-dict translated fields.
-        m2m:              M2MConfig list for M2M relations via through tables.
-        lookup_field:     Single field for update_or_create lookup (default: "external_id").
-        api_id_key:       Primary identifier key in the API object (default: "@id").
-        item_filter:      Optional predicate - return False to skip an item.
-    """
-
-    field_map: Dict[str, Optional[str]] = field(default_factory=dict)
-    value_transforms: Dict[str, Callable[[Any], Any]] = field(default_factory=dict)
-    fk_resolvers: Dict[str, Callable[[Any], Optional[Any]]] = field(default_factory=dict)
-    translations: List[TranslationConfig] = field(default_factory=list)
-    m2m: List[M2MConfig] = field(default_factory=list)
-    lookup_field: str = "external_id"
-    api_id_key: str = "@id"
-    item_filter: Optional[Callable[[Mapping[str, Any]], bool]] = None
-
-
-# ---------------------------------------------------------------------------
-# HTTP session with retry + exponential backoff
-# ---------------------------------------------------------------------------
+normalize_url = _normalize.normalize_url
+normalize_performer_type = _normalize.normalize_performer_type
+clean_string = _normalize.clean_string
+clean_vendor_id = _normalize.clean_vendor_id
+nee_ja_to_bool = _normalize.nee_ja_to_bool
 
 
 def _get_api_key_or_raise() -> str:
-    api_key = os.getenv("VIERNULVIER_API_KEY")
-    if not api_key:
-        raise ScraperError("VIERNULVIER_API_KEY environment variable is not set")
-    return api_key
+    return _http.get_api_key_or_raise()
 
 
 def _build_session() -> requests.Session:
-    """Create a requests.Session with connection pooling and shared auth headers."""
-    session = requests.Session()
-    session.headers.update(
-        {
-            "X-AUTH-TOKEN": _get_api_key_or_raise(),
-            "accept": "application/ld+json",
-            "User-Agent": random.choice(USER_AGENT_POOL),
-        }
-    )
-    adapter = requests.adapters.HTTPAdapter(
-        pool_connections=MAX_CONCURRENT_PAGES,
-        pool_maxsize=MAX_CONCURRENT_PAGES * 2,
-    )
-    session.mount("https://", adapter)
-    session.mount("http://", adapter)
-    return session
+    return _http.build_session()
 
 
 def _backoff_seconds(attempt: int) -> float:
-    """Exponential backoff with full jitter, capped at RETRY_BACKOFF_MAX."""
-    return min(
-        RETRY_BACKOFF_BASE * (2**attempt) + random.uniform(0, 1),
-        RETRY_BACKOFF_MAX,
-    )
+    return _http.backoff_seconds(attempt, backoff_base=RETRY_BACKOFF_BASE, backoff_max=RETRY_BACKOFF_MAX)
 
 
 def _parse_retry_after(response: requests.Response) -> Optional[int]:
-    """Parse the Retry-After response header as an integer number of seconds."""
-    header = response.headers.get("Retry-After")
-    if header:
-        try:
-            return int(header)
-        except ValueError:
-            pass
-    return None
+    return _http.parse_retry_after(response)
 
 
 def _fetch_with_retry(
@@ -259,129 +95,22 @@ def _fetch_with_retry(
     params: Optional[Dict[str, str]] = None,
     etag: Optional[str] = None,
 ) -> Tuple[Optional[Any], Optional[str]]:
-    """Fetch one URL with retry + exponential backoff + jitter.
-
-    Handles:
-    - HTTP 304 Not Modified  -> returns (None, cached_etag)
-    - HTTP 429 Too Many Requests -> respects Retry-After header
-    - HTTP 5xx server errors -> retries with backoff
-    - ConnectionError / Timeout -> retries with backoff
-    - Other transport errors -> raises immediately (not retryable)
-
-    Returns:
-        (data, new_etag). data is None when the server returns 304.
-    """
-    headers: Dict[str, str] = {}
-    if etag:
-        headers["If-None-Match"] = etag
-
-    for attempt in range(MAX_RETRIES + 1):
-        try:
-            response = session.get(url, params=params, headers=headers, timeout=DEFAULT_TIMEOUT)
-        except requests.ConnectionError as exc:
-            if attempt == MAX_RETRIES:
-                raise ScraperError(f"Connection failed after {MAX_RETRIES} attempts: {url}") from exc
-            wait = _backoff_seconds(attempt)
-            logger.warning("ConnectionError - retry %d in %.1fs: %s", attempt + 1, wait, url)
-            time.sleep(wait)
-            continue
-        except requests.Timeout as exc:
-            if attempt == MAX_RETRIES:
-                raise ScraperError(f"Request timed out after {MAX_RETRIES} attempts: {url}") from exc
-            wait = _backoff_seconds(attempt)
-            logger.warning("Timeout - retry %d in %.1fs: %s", attempt + 1, wait, url)
-            time.sleep(wait)
-            continue
-        except requests.RequestException as exc:
-            # SSL errors, invalid URL, etc. - not worth retrying
-            raise ScraperError(f"Request failed: {url}") from exc
-
-        if response.status_code == 304:
-            logger.debug("304 Not Modified - page unchanged: %s", url)
-            return None, etag
-
-        if response.status_code == 429:
-            retry_after = _parse_retry_after(response)
-            if attempt == MAX_RETRIES:
-                raise RateLimitError(retry_after)
-            wait = float(retry_after) if retry_after else _backoff_seconds(attempt)
-            logger.warning("HTTP 429 - waiting %.1fs before retry %d: %s", wait, attempt + 1, url)
-            time.sleep(wait)
-            continue
-
-        if response.status_code in RETRY_STATUS_CODES:
-            if attempt == MAX_RETRIES:
-                raise ScraperError(f"HTTP {response.status_code} after {MAX_RETRIES} attempts: {url}")
-            wait = _backoff_seconds(attempt)
-            logger.warning(
-                "HTTP %d - retry %d in %.1fs: %s",
-                response.status_code,
-                attempt + 1,
-                wait,
-                url,
-            )
-            time.sleep(wait)
-            continue
-
-        if not response.ok:
-            raise ScraperError(f"API error: {response.status_code} - {url}")
-
-        try:
-            data = response.json()
-        except ValueError as exc:
-            raise ScraperError(f"Invalid JSON from API: {url}") from exc
-
-        if data is None:
-            raise ScraperError(f"API returned null payload: {url}")
-
-        if isinstance(data, dict) and data.get("@context") == ERROR_CONTEXT_PATH:
-            status = data.get("status", "unknown")
-            detail = data.get("detail", "no detail provided")
-            raise ScraperError(f"API error response: status={status}, detail={detail}")
-
-        new_etag = response.headers.get("ETag")
-        return data, new_etag
-
-    raise ScraperError(f"Retries exhausted for: {url}")  # unreachable
-
-
-# ---------------------------------------------------------------------------
-# Pagination discovery
-# ---------------------------------------------------------------------------
+    return _http.fetch_with_retry(
+        session,
+        url,
+        params=params,
+        etag=etag,
+        max_retries=MAX_RETRIES,
+        timeout=DEFAULT_TIMEOUT,
+        retry_status_codes=RETRY_STATUS_CODES,
+        sleep_fn=time.sleep,
+        backoff_fn=_backoff_seconds,
+        parse_retry_after_fn=_parse_retry_after,
+    )
 
 
 def _discover_extra_pages(data: dict) -> List[str]:
-    """Discover page URLs beyond page 1 from the hydra:view metadata.
-
-    Primary strategy: parse the "last" URL for the total page count and
-    generate all intermediate page URLs by substituting the page number.
-    This allows full concurrent fetching.
-
-    Falls back to an empty list if the "last" URL does not contain ?page=N,
-    which signals the caller to fall back to sequential "next" link following.
-
-    Returns a list of page URLs for pages 2..N in order.
-    """
-    view = data.get("view") or {}
-    last_url = view.get("last", "")
-
-    match = re.search(r"[?&]page=(\d+)", last_url)
-    if not match:
-        return []  # Caller will use sequential fallback
-
-    total_pages = int(match.group(1))
-    pages = []
-    for page_num in range(2, total_pages + 1):
-        page_url = re.sub(r"([?&])page=\d+", rf"\g<1>page={page_num}", last_url)
-        if not page_url.startswith("http"):
-            page_url = urljoin(BASE_DOMAIN, page_url)
-        pages.append(page_url)
-    return pages
-
-
-# ---------------------------------------------------------------------------
-# Main fetch function
-# ---------------------------------------------------------------------------
+    return _http.discover_extra_pages(data, base_domain=BASE_DOMAIN)
 
 
 def fetch_viernulvier(
@@ -389,403 +118,32 @@ def fetch_viernulvier(
     params: Optional[Dict[str, str]] = None,
     etag_cache: Optional[Dict[str, str]] = None,
 ) -> List[Any]:
-    """Fetch all items from an API endpoint, following pagination automatically.
-
-    Two pagination strategies:
-    1. Concurrent (preferred): if the "last" URL has a ?page=N pattern, derive
-       all page URLs upfront and fetch them in parallel.
-    2. Sequential fallback: follow "next" links one by one (slower but handles
-       non-standard pagination).
-
-    Args:
-        endpoint:    Relative API path (e.g. "/productions").
-        params:      Optional query parameters, applied only to page 1.
-        etag_cache:  Optional dict of url -> ETag. Pass the same instance across
-                     calls to benefit from 304 Not Modified responses.
-
-    Returns:
-        All items across all pages in original order.
-    """
-    parsed = urlparse(endpoint)
-    if parsed.scheme or parsed.netloc:
-        raise ScraperError(f"endpoint must be a relative path, got: {endpoint!r}")
-
-    url = urljoin(BASE_URL + "/", endpoint.lstrip("/"))
-    etag_cache = etag_cache if etag_cache is not None else {}
-    session = _build_session()
-
-    logger.info("Fetching: %s", url)
-    data, new_etag = _fetch_with_retry(session, url, params=params, etag=etag_cache.get(url))
-    if new_etag:
-        etag_cache[url] = new_etag
-
-    if data is None:
-        logger.info("304 Not Modified for %s - nothing to sync.", url)
-        return []
-
-    if isinstance(data, list):
-        return data
-
-    if not isinstance(data, dict):
-        raise ScraperError(f"Unexpected payload type: {type(data)}")
-
-    members = data.get("member", [])
-    if not members and "@context" in data:
-        members = [data]  # Single-object response wrapped in hydra context
-    all_items: List[Any] = list(members)
-
-    total_items = data.get("totalItems") or data.get("hydra:totalItems")
-    extra_pages = _discover_extra_pages(data)
-
-    if total_items:
-        logger.info(
-            "API reports %d total items - %d additional pages to fetch",
-            total_items,
-            len(extra_pages),
-        )
-
-    # --- Strategy 1: concurrent page fetch (all URLs known upfront) ---
-    if extra_pages:
-        page_results: Dict[str, List[Any]] = {}
-
-        with ThreadPoolExecutor(max_workers=MAX_CONCURRENT_PAGES) as executor:
-            future_to_url = {
-                executor.submit(
-                    _fetch_with_retry,
-                    session,
-                    page_url,
-                    None,  # query params only on page 1
-                    etag_cache.get(page_url),
-                ): page_url
-                for page_url in extra_pages
-            }
-            for future in as_completed(future_to_url):
-                page_url = future_to_url[future]
-                try:
-                    page_data, page_etag = future.result()
-                    if page_etag:
-                        etag_cache[page_url] = page_etag
-                    if page_data is None:
-                        continue  # 304
-                    page_results[page_url] = page_data.get("member", []) if isinstance(page_data, dict) else page_data
-                except ScraperError as exc:
-                    logger.error("Page fetch failed for %s: %s", page_url, exc)
-
-        # Merge in original page order (as_completed is unordered)
-        for page_url in extra_pages:
-            all_items.extend(page_results.get(page_url, []))
-
-    # --- Strategy 2: sequential "next" link fallback ---
-    else:
-        view = data.get("view") or {}
-        next_raw = view.get("next")
-        current_url: Optional[str] = (
-            (next_raw if next_raw.startswith("http") else urljoin(BASE_DOMAIN, next_raw)) if next_raw else None
-        )
-        while current_url:
-            page_data, page_etag = _fetch_with_retry(session, current_url, etag=etag_cache.get(current_url))
-            if page_etag:
-                etag_cache[current_url] = page_etag
-            if page_data is None:
-                break
-            if isinstance(page_data, dict):
-                all_items.extend(page_data.get("member", []))
-                next_view = page_data.get("view") or {}
-                next_raw = next_view.get("next")
-                current_url = (
-                    (next_raw if next_raw.startswith("http") else urljoin(BASE_DOMAIN, next_raw)) if next_raw else None
-                )
-            else:
-                all_items.extend(page_data if isinstance(page_data, list) else [])
-                break
-
-    logger.info("Fetched %d items from %s", len(all_items), endpoint)
-    return all_items
-
-
-# ---------------------------------------------------------------------------
-# Value normalisation helpers (exported for use in management commands)
-# ---------------------------------------------------------------------------
-
-_url_validator = URLValidator()
-
-
-def normalize_url(value: Any) -> str:
-    """Return the value only if it is a valid URL, otherwise return empty string."""
-    if value is None:
-        return ""
-    s = str(value).strip()
-    if s.lower() in _EMPTY_URL_VALUES:
-        return ""
-    try:
-        _url_validator(s)
-        return s
-    except ValidationError:
-        return ""
-
-
-def normalize_performer_type(value: Any) -> str:
-    """Map API performer_type values to internal enum values."""
-    _mapping = {"person": "solo"}
-    s = str(value).strip().lower() if value else ""
-    return _mapping.get(s, s)
-
-
-def clean_string(value: Any) -> str:
-    """Strip surrounding whitespace and remove non-printable control characters.
-
-    Keeps tab (\\t), line feed (\\n), and carriage return (\\r) intact.
-    Removes null bytes and other ASCII control characters that would corrupt
-    the database or cause JSON serialisation errors.
-    """
-    if value is None:
-        return ""
-    s = str(value).strip()
-    s = re.sub(r"[\x00-\x08\x0b\x0c\x0e-\x1f\x7f]", "", s)
-    return s
-
-
-def clean_vendor_id(value: Any) -> Optional[str]:
-    """
-    Transform vendor_id: return None if empty or HTML-like, else return as string.
-    """
-    if value is None:
-        return None
-    s = str(value).strip()
-    if not s or s.startswith("<i"):
-        return None
-    return s
-
-
-def nee_ja_to_bool(value: Any) -> bool:
-    """Coerce Dutch/English truthy strings and integers to Python bool.
-
-    Accepts: "ja"/"nee", "true"/"false", "yes"/"no", "1"/"0", bool, int.
-    """
-    if isinstance(value, bool):
-        return value
-    if isinstance(value, int):
-        return bool(value)
-    if isinstance(value, str):
-        v = value.strip().lower()
-        if v in {"ja", "true", "1", "yes"}:
-            return True
-        if v in {"nee", "false", "0", "", "no"}:
-            return False
-    return bool(value)
-
-
-def _camel_to_snake(value: str) -> str:
-    """Convert camelCase API key to snake_case model field name."""
-    return re.sub(r"(?<!^)(?=[A-Z])", "_", value).lower()
+    return _http.fetch_viernulvier_impl(
+        endpoint=endpoint,
+        params=params,
+        etag_cache=etag_cache,
+        base_url=BASE_URL,
+        base_domain=BASE_DOMAIN,
+        build_session_fn=_build_session,
+        fetch_with_retry_fn=_fetch_with_retry,
+        discover_extra_pages_fn=_discover_extra_pages,
+    )
 
 
 def _parse_field_value(model_field: models.Field, value: Any) -> Any:
-    """Coerce an API value to the correct Python type for the given model field.
-
-    Supported field types:
-    - CharField, TextField   -> clean_string, respect max_length
-    - URLField               -> normalize_url
-    - BooleanField           -> nee_ja_to_bool
-    - DecimalField           -> Decimal (avoids float rounding errors for prices)
-    - IntegerField           -> int
-    - FloatField             -> float
-    - DateTimeField          -> parse_datetime (with broken-date repair)
-    - DateField              -> parse_date
-    - All other types        -> pass through unchanged
-    """
-    if value is None:
-        return None
-
-    if isinstance(model_field, (models.TextField, models.CharField)):
-        cleaned = clean_string(value)
-        max_len = getattr(model_field, "max_length", None)
-        if max_len and len(cleaned) > max_len:
-            logger.debug(
-                "Field '%s' truncated: %d -> %d chars",
-                model_field.name,
-                len(cleaned),
-                max_len,
-            )
-            cleaned = cleaned[:max_len]
-        return cleaned
-
-    if isinstance(model_field, models.URLField):
-        return normalize_url(value)
-
-    if isinstance(model_field, models.BooleanField):
-        return nee_ja_to_bool(value)
-
-    if isinstance(model_field, models.DecimalField):
-        # Use Decimal - never float - for monetary values to avoid rounding errors.
-        try:
-            return Decimal(str(value))
-        except (InvalidOperation, ValueError, TypeError):
-            logger.warning("Cannot convert '%s' to Decimal for field '%s'", value, model_field.name)
-            return None
-
-    if isinstance(model_field, models.IntegerField):
-        try:
-            return int(value)
-        except (ValueError, TypeError):
-            logger.warning("Cannot convert '%s' to int for field '%s'", value, model_field.name)
-            return None
-
-    if isinstance(model_field, models.FloatField):
-        try:
-            return float(value)
-        except (ValueError, TypeError):
-            logger.warning("Cannot convert '%s' to float for field '%s'", value, model_field.name)
-            return None
-
-    if isinstance(model_field, models.DateTimeField) and isinstance(value, str):
-        v = value
-        # Repair two broken date formats observed in this API
-        if v.startswith("-"):
-            v = v[1:]
-        if v[:4] == "0000":
-            v = "1970" + v[4:]
-        parsed = parse_datetime(v)
-        if parsed is None:
-            logger.debug("Cannot parse datetime '%s' for field '%s'", value, model_field.name)
-        return parsed
-
-    if isinstance(model_field, models.DateField) and isinstance(value, str):
-        return parse_date(value)
-
-    return value
-
-
-# ---------------------------------------------------------------------------
-# FK cache - eliminates N+1 queries
-# ---------------------------------------------------------------------------
-
-
-class FKCache:
-    """In-memory cache of (model, external_id) -> pk.
-
-    On first use of a model, loads all known external_id -> pk pairs with a
-    single bulk queryset (warmup). Subsequent FK resolutions are O(1) dict
-    lookups with a single DB query only on a true cache miss.
-
-    Warmup failures are recorded so that the failed model is not re-queried
-    on every lookup (avoids repeated broken queries for models without
-    external_id).
-
-    Thread-safety: not thread-safe. Intended for use in sync loops only.
-    """
-
-    def __init__(self) -> None:
-        self._cache: Dict[Tuple[Type[models.Model], str], Any] = {}
-        # True = warmed up successfully, False = warmup failed
-        self._loaded: Dict[Type[models.Model], bool] = {}
-
-    def warmup(self, model: Type[models.Model]) -> None:
-        """Pre-load all external_id -> pk mappings for a model in one query."""
-        if model in self._loaded:
-            return
-        try:
-            count = 0
-            for ext_id, pk in model.objects.values_list("external_id", "pk").iterator():
-                self._cache[(model, str(ext_id))] = pk
-                count += 1
-            self._loaded[model] = True
-            logger.debug("FK cache warmed: %s (%d entries)", model.__name__, count)
-        except Exception:
-            # Mark as attempted (False) so we do not keep retrying this broken query.
-            self._loaded[model] = False
-            logger.warning(
-                "FK cache warmup failed for %s (no external_id field?)",
-                model.__name__,
-                exc_info=True,
-            )
-
-    def get(self, model: Type[models.Model], ext_id: str) -> Optional[Any]:
-        self.warmup(model)
-        return self._cache.get((model, str(ext_id)))
-
-    def set(self, model: Type[models.Model], ext_id: str, pk: Any) -> None:
-        """Insert or update a cache entry after creating a new object."""
-        self._cache[(model, str(ext_id))] = pk
-
-
-# ---------------------------------------------------------------------------
-# Helper utilities
-# ---------------------------------------------------------------------------
+    return _normalize.parse_field_value(model_field, value)
 
 
 def _extract_external_id_from_url(raw: Any) -> Optional[str]:
-    """Extract an external ID string from a URL, embedded dict, or integer.
-
-    Examples:
-    - "https://www.viernulvier.gent/api/v1/halls/42" -> full URL string
-    - {"@id": "/api/v1/productions/5", ...}          -> "/api/v1/productions/5"
-    - 42                                              -> "42"
-    """
-    if raw is None:
-        return None
-    if isinstance(raw, int):
-        return str(raw)
-    if isinstance(raw, dict):
-        raw = raw.get("@id") or raw.get("external_id") or raw.get("id")
-        if raw is None:
-            return None
-    if isinstance(raw, str):
-        return raw.strip() or None
-    return None
+    return _relations.extract_external_id_from_url(raw)
 
 
-def _resolve_fk(
-    model_field: models.Field,
-    raw_value: Any,
-    fk_cache: FKCache,
-) -> Optional[Any]:
-    """Resolve a FK raw value (URL or embedded dict) to a database primary key.
-
-    Uses the FK cache for fast lookups. On cache miss, queries the DB once
-    and populates the cache so future calls for the same object are free.
-    """
-    related_model = model_field.remote_field.model
-    ext_id = _extract_external_id_from_url(raw_value)
-    if ext_id is None:
-        return None
-
-    pk = fk_cache.get(related_model, ext_id)
-    if pk is not None:
-        return pk
-
-    try:
-        pk = related_model.objects.values_list("pk", flat=True).get(external_id=ext_id)
-        fk_cache.set(related_model, ext_id, pk)
-        return pk
-    except related_model.DoesNotExist:
-        logger.warning(
-            "FK not found: %s.external_id=%r - sync related models first.",
-            related_model.__name__,
-            ext_id,
-        )
-        return None
-    except Exception:
-        logger.exception("Error resolving FK %s external_id=%r", related_model.__name__, ext_id)
-        return None
+def _resolve_fk(model_field: models.Field, raw_value: Any, fk_cache: FKCache) -> Optional[Any]:
+    return _relations.resolve_fk(model_field, raw_value, fk_cache)
 
 
-def _extract_lookup_value(
-    item: Mapping[str, Any],
-    config: ModelSyncConfig,
-) -> Optional[str]:
-    """Extract the primary lookup value (usually the API @id) from an item."""
-    raw = item.get(config.api_id_key) or item.get("external_id") or item.get("id")
-    if raw is None:
-        return None
-    if isinstance(raw, dict):
-        raw = raw.get("@id") or raw.get("external_id") or raw.get("id")
-    return str(raw).strip() if raw is not None else None
-
-
-# ---------------------------------------------------------------------------
-# Build defaults dict: API item -> Django model kwargs
-# ---------------------------------------------------------------------------
+def _extract_lookup_value(item: Mapping[str, Any], config: ModelSyncConfig) -> Optional[str]:
+    return _relations.extract_lookup_value(item, config)
 
 
 def _build_defaults(
@@ -794,76 +152,7 @@ def _build_defaults(
     config: ModelSyncConfig,
     fk_cache: FKCache,
 ) -> Dict[str, Any]:
-    """Convert an API item to a defaults dict for update_or_create.
-
-    Pass 1: process explicit field_map entries (highest priority).
-    Pass 2: auto-map remaining keys via camelCase -> snake_case conversion.
-
-    Flat dicts (translations) and lists (M2M) are skipped here.
-    """
-    defaults: Dict[str, Any] = {}
-    explicitly_mapped = set(config.field_map.keys())
-
-    def _apply(model_field: models.Field, raw_value: Any, field_name: str) -> None:
-        if model_field.is_relation and model_field.many_to_one:
-            custom = config.fk_resolvers.get(field_name)
-            pk = custom(raw_value) if custom else _resolve_fk(model_field, raw_value, fk_cache)
-            if pk is not None:
-                defaults[f"{model_field.name}_id"] = pk
-        else:
-            converted = _parse_field_value(model_field, raw_value)
-            transform = config.value_transforms.get(field_name)
-            if transform:
-                converted = transform(converted)
-            if converted is not None:
-                defaults[model_field.name] = converted
-
-    # Pass 1: explicit field_map
-    for api_key, model_field_name in config.field_map.items():
-        if model_field_name is None:
-            continue
-        raw_value = item.get(api_key)
-        if raw_value is None:
-            continue
-        try:
-            model_field = model._meta.get_field(model_field_name)
-        except FieldDoesNotExist:
-            logger.warning("Field '%s' does not exist on %s", model_field_name, model.__name__)
-            continue
-        if not isinstance(model_field, models.Field) or model_field.primary_key:
-            continue
-        if isinstance(raw_value, dict) and not model_field.is_relation:
-            continue  # flat dict -> handled by translation sync
-        _apply(model_field, raw_value, model_field_name)
-
-    # Pass 2: auto-map unmapped keys
-    for api_key, raw_value in item.items():
-        if api_key in explicitly_mapped or api_key.startswith("@") or raw_value is None:
-            continue
-        if isinstance(raw_value, (list, dict)):
-            continue
-
-        snake_key = _camel_to_snake(api_key)
-        model_field = None
-        for candidate in (api_key, snake_key):
-            try:
-                f = model._meta.get_field(candidate)
-                if isinstance(f, models.Field) and not f.primary_key:
-                    model_field = f
-                    break
-            except FieldDoesNotExist:
-                pass
-
-        if model_field is None:
-            continue
-        _apply(model_field, raw_value, model_field.name)
-
-    return defaults
-
-
-# ---------------------------------------------------------------------------
-# Translation sync - batched per language to minimise query count
-# ---------------------------------------------------------------------------
+    return _relations.build_defaults(model, item, config, fk_cache, resolve_fk_fn=_resolve_fk)
 
 
 def _sync_all_translations(
@@ -871,172 +160,11 @@ def _sync_all_translations(
     item: Mapping[str, Any],
     translation_configs: List[TranslationConfig],
 ) -> None:
-    """Sync all translated fields for one parent object.
-
-    Batching: configs that share the same (translation model, parent_fk,
-    language_fk) are grouped together. For each language code, all field
-    values from every config in the group are merged, then saved with a
-    single update_or_create call.
-
-    Without batching: 13 fields x 3 languages = 39 queries per production.
-    With batching:                               3 queries per production.
-
-    Configs targeting different translation models are handled independently.
-    """
-    if not translation_configs:
-        return
-
-    groups = defaultdict(list)
-    for cfg in translation_configs:
-        groups[(cfg.model, cfg.parent_fk, cfg.language_fk)].append(cfg)
-
-    for (trans_model, parent_fk, language_fk), cfgs in groups.items():
-        # Collect every language code present across all field dicts
-        all_languages: Set[str] = set()
-        for cfg in cfgs:
-            raw_dict = item.get(cfg.api_key)
-            if isinstance(raw_dict, dict):
-                all_languages.update(raw_dict.keys())
-
-        for lang_code in all_languages:
-            if not lang_code:
-                continue
-
-            field_updates: Dict[str, Any] = {}
-
-            for cfg in cfgs:
-                raw_dict = item.get(cfg.api_key)
-                if not isinstance(raw_dict, dict):
-                    continue
-                raw_value = raw_dict.get(lang_code)
-                if raw_value is None:
-                    continue
-
-                try:
-                    model_field = trans_model._meta.get_field(cfg.flat_field)
-                except FieldDoesNotExist:
-                    logger.warning(
-                        "Translation field '%s' not found on %s",
-                        cfg.flat_field,
-                        trans_model.__name__,
-                    )
-                    continue
-
-                if raw_value == "" and not getattr(model_field, "blank", True):
-                    continue
-
-                converted = _parse_field_value(model_field, raw_value)
-                transform = cfg.value_transforms.get(cfg.flat_field)
-                if transform:
-                    converted = transform(converted)
-                if converted is not None:
-                    field_updates[cfg.flat_field] = converted
-
-            if not field_updates:
-                continue
-
-            try:
-                trans_model.objects.update_or_create(
-                    **{parent_fk: parent_obj, language_fk: lang_code},
-                    defaults=field_updates,
-                )
-            except Exception:
-                logger.exception(
-                    "Error syncing %s translations for %s pk=%s lang=%s fields=%s",
-                    trans_model.__name__,
-                    parent_obj.__class__.__name__,
-                    parent_obj.pk,
-                    lang_code,
-                    list(field_updates.keys()),
-                )
+    _relations.sync_all_translations(parent_obj, item, translation_configs)
 
 
-# ---------------------------------------------------------------------------
-# M2M sync via through table
-# ---------------------------------------------------------------------------
-
-
-def _sync_m2m(
-    parent_obj: models.Model,
-    item: Mapping[str, Any],
-    m2m_config: M2MConfig,
-    fk_cache: FKCache,
-) -> None:
-    """Sync one M2M relation via its through table using bulk_create.
-
-    Deletes all existing through-table rows for this parent, then recreates
-    them from the current API data. Uses bulk_create for performance with an
-    individual-save fallback if bulk_create fails.
-    """
-    raw_list = item.get(m2m_config.api_key)
-    if not isinstance(raw_list, list):
-        return
-
-    through_model = m2m_config.through_model
-    related_model = m2m_config.related_model
-
-    fk_cache.warmup(related_model)
-    through_model.objects.filter(**{m2m_config.parent_fk: parent_obj}).delete()
-
-    to_create = []
-    for position, raw_item in enumerate(raw_list):
-        ext_id = _extract_external_id_from_url(raw_item)
-        if not ext_id:
-            continue
-
-        pk = fk_cache.get(related_model, ext_id)
-        if pk is None:
-            try:
-                obj = related_model.objects.get(**{m2m_config.related_lookup_field: ext_id})
-                fk_cache.set(related_model, ext_id, obj.pk)
-                pk = obj.pk
-            except related_model.DoesNotExist:
-                logger.warning(
-                    "%s with %s=%r not found - sync related models first.",
-                    related_model.__name__,
-                    m2m_config.related_lookup_field,
-                    ext_id,
-                )
-                continue
-
-        through_kwargs: Dict[str, Any] = {
-            m2m_config.parent_fk: parent_obj,
-            m2m_config.related_fk: related_model(pk=pk),
-        }
-        for api_field, through_field in m2m_config.extra_fields.items():
-            val = raw_item.get(api_field) if isinstance(raw_item, dict) else None
-            if val is None and api_field == "position":
-                val = position
-            if val is not None:
-                through_kwargs[through_field] = val
-
-        to_create.append(through_model(**through_kwargs))
-
-    if not to_create:
-        return
-
-    try:
-        through_model.objects.bulk_create(to_create, ignore_conflicts=True)
-    except Exception:
-        logger.warning(
-            "bulk_create failed for %s - falling back to individual saves",
-            through_model.__name__,
-        )
-        for obj in to_create:
-            try:
-                obj.save()
-            except Exception:
-                logger.exception(
-                    "Error creating %s for %s pk=%s",
-                    through_model.__name__,
-                    parent_obj.__class__.__name__,
-                    parent_obj.pk,
-                )
-
-
-# ---------------------------------------------------------------------------
-# Main sync function
-# ---------------------------------------------------------------------------
+def _sync_m2m(parent_obj: models.Model, item: Mapping[str, Any], m2m_config: M2MConfig, fk_cache: FKCache) -> None:
+    _relations.sync_m2m(parent_obj, item, m2m_config, fk_cache)
 
 
 def sync_viernulvier(
@@ -1048,183 +176,37 @@ def sync_viernulvier(
     etag_cache: Optional[Dict[str, str]] = None,
     on_progress: Optional[Callable[[int, int], None]] = None,
 ) -> int:
-    """Fetch Viernulvier API data and persist it to a Django model.
-
-    Per-item savepoints ensure one bad record never aborts the full batch.
-    The ImportLog is updated with totals and error summaries at the end.
-
-    Args:
-        model:        Django model class.
-        config:       ModelSyncConfig describing field mapping, translations, M2M.
-        endpoint:     Relative API path (e.g. "/productions").
-        params:       Optional query parameters forwarded to the API.
-        dry_run:      When True, fetch and parse but do not write to the DB.
-        etag_cache:   Shared ETag dict; pass the same instance across all sync
-                      steps so unchanged endpoints are skipped automatically.
-        on_progress:  Callback(saved, total) invoked after each successful save.
-
-    Returns:
-        Number of records saved (created + updated). Always 0 in dry_run mode.
-    """
-    from apps.import_log.models import ImportLog
-
-    source = f"viernulvier:{endpoint}"
-    if params:
-        params_str = ",".join(f"{k}={v}" for k, v in sorted(params.items()))
-        # Guard against max_length violations on ImportLog.source
-        suffix = f"?{params_str}"
-        source = source[: 200 - len(suffix)] + suffix
-
-    import_log = ImportLog.objects.create(
-        source=source,
-        status=ImportLog.Status.IN_PROGRESS,
-        started_at=timezone.now(),
+    return _sync.sync_viernulvier_impl(
+        model,
+        config,
+        fetch_fn=fetch_viernulvier,
+        build_defaults_fn=_build_defaults,
+        sync_translations_fn=_sync_all_translations,
+        sync_m2m_fn=_sync_m2m,
+        extract_lookup_value_fn=_extract_lookup_value,
+        fk_cache_cls=FKCache,
+        transaction_module=transaction,
+        timezone_module=timezone,
+        endpoint=endpoint,
+        params=params,
+        dry_run=dry_run,
+        etag_cache=etag_cache,
+        on_progress=on_progress,
     )
 
-    try:
-        items = fetch_viernulvier(endpoint=endpoint, params=params, etag_cache=etag_cache)
-    except Exception as exc:
-        import_log.status = ImportLog.Status.FAILED
-        import_log.finished_at = timezone.now()
-        import_log.error_message = str(exc)
-        import_log.save()
-        raise
 
-    if not items:
-        import_log.status = ImportLog.Status.SUCCESS
-        import_log.records_total = 0
-        import_log.records_imported = 0
-        import_log.records_failed = 0
-        import_log.finished_at = timezone.now()
-        import_log.save()
-        return 0
+def _derive_crop_filename(crop_name: str, item_external_id: str, image_url: str) -> str:
+    return _media.derive_crop_filename(crop_name, item_external_id, image_url)
 
-    # Pre-warm FK cache for all M2M related models
-    fk_cache = FKCache()
-    for m2m_cfg in config.m2m:
-        fk_cache.warmup(m2m_cfg.related_model)
 
-    saved = 0
-    errors = 0
-    error_messages: List[str] = []  # capped at MAX_ERROR_MESSAGES
-    seen: Set[str] = set()
-    total = len(items)
-
-    def _record_error(msg: str) -> None:
-        nonlocal errors
-        errors += 1
-        if len(error_messages) < MAX_ERROR_MESSAGES:
-            error_messages.append(msg)
-
-    for item in items:
-        if not isinstance(item, dict):
-            _record_error(f"Item is not a dict: {item!r}")
-            logger.error("Item is not a dict: %r", item)
-            continue
-
-        lookup_value = _extract_lookup_value(item, config)
-
-        if config.item_filter and not config.item_filter(item):
-            logger.debug("Item filtered out: %s", lookup_value)
-            continue
-
-        if lookup_value is None:
-            msg = f"Missing '{config.api_id_key}' in item: {str(item)[:200]}"
-            logger.warning(msg)
-            _record_error(msg)
-            continue
-
-        if lookup_value in seen:
-            logger.warning("Duplicate item skipped: %s", lookup_value)
-            continue
-        seen.add(lookup_value)
-
-        if dry_run:
-            defaults = _build_defaults(model, item, config, fk_cache)
-            logger.info("[DRY RUN] Would save %s (%d fields)", lookup_value, len(defaults))
-            saved += 1
-            if on_progress:
-                on_progress(saved, total)
-            continue
-
-        sid = transaction.savepoint()
-        try:
-            defaults = _build_defaults(model, item, config, fk_cache)
-
-            lookup_kwargs = {config.lookup_field: lookup_value}
-
-            obj, created = model.objects.update_or_create(
-                **lookup_kwargs,
-                defaults=defaults,
-            )
-
-            # Cache the new object so downstream FK lookups within this batch are free
-            if hasattr(obj, "external_id"):
-                fk_cache.set(model, str(obj.external_id), obj.pk)
-
-            # Batched translation sync: one query per language, not per field
-            _sync_all_translations(obj, item, config.translations)
-
-            for m2m_cfg in config.m2m:
-                _sync_m2m(obj, item, m2m_cfg, fk_cache)
-
-            transaction.savepoint_commit(sid)
-            saved += 1
-
-            if on_progress:
-                on_progress(saved, total)
-
-            logger.debug(
-                "%s %s: %s",
-                "Created" if created else "Updated",
-                model.__name__,
-                lookup_value,
-            )
-
-        except ValidationError as exc:
-            transaction.savepoint_rollback(sid)
-            msgs = [f"{f}: {err}" if f != "__all__" else err for f, errs in exc.message_dict.items() for err in errs]
-            _record_error(f"Validation error for {lookup_value}: {'; '.join(msgs)}")
-            logger.error("Validation error for %s: %s", lookup_value, "; ".join(msgs))
-
-        except (IntegrityError, DatabaseError, FieldError):
-            transaction.savepoint_rollback(sid)
-            exc_type, exc_value, _ = sys.exc_info()
-            msg = f"Database error for {lookup_value}: {exc_type.__name__}: {exc_value}"
-            _record_error(msg)
-            logger.error(msg, exc_info=True)
-
-        except Exception:
-            transaction.savepoint_rollback(sid)
-            exc_type, exc_value, _ = sys.exc_info()
-            msg = f"Unexpected error for {lookup_value}: {exc_type.__name__}: {exc_value}"
-            _record_error(msg)
-            logger.error(msg, exc_info=True)
-
-    # Finalise import log
-    truncation_note = f" (showing first {MAX_ERROR_MESSAGES} of {errors})" if errors > MAX_ERROR_MESSAGES else ""
-    import_log.records_total = total
-    import_log.records_imported = saved
-    import_log.records_failed = errors
-    import_log.finished_at = timezone.now()
-
-    if errors == 0:
-        import_log.status = ImportLog.Status.SUCCESS
-    elif saved > 0:
-        import_log.status = ImportLog.Status.PARTIAL_SUCCESS
-        import_log.error_message = f"{errors} records failed{truncation_note}: {', '.join(error_messages)}"
-    else:
-        import_log.status = ImportLog.Status.FAILED
-        import_log.error_message = f"All {errors} records failed{truncation_note}: {', '.join(error_messages)}"
-
-    import_log.save()
-    logger.info(
-        "Sync complete: saved=%d, errors=%d%s",
-        saved,
-        errors,
-        " [DRY RUN]" if dry_run else "",
+def _download_image(session: requests.Session, url: str) -> Optional[bytes]:
+    return _media.download_image(
+        session,
+        url,
+        max_retries=MAX_RETRIES,
+        sleep_fn=time.sleep,
+        backoff_fn=_backoff_seconds,
     )
-    return saved
 
 
 def sync_media_item_gallery_links(
@@ -1233,286 +215,15 @@ def sync_media_item_gallery_links(
     on_progress: Optional[Callable[[int, int], None]] = None,
     params: Optional[Dict[str, str]] = None,
 ) -> int:
-    """Link MediaItems to MediaGalleries using gallery payload `items` links.
-
-    The Peppered `/media/items` payload often has `gallery: null`. The
-    `/media/galleries` payload is the authoritative source because it embeds the
-    media item links. This sync resolves those links and updates
-    `MediaItem.gallery` (and `position`) in bulk.
-
-    Returns the number of MediaItem rows changed for compatibility fields
-    (`gallery` / `position`). Through-table writes are tracked separately for
-    observability but are not part of the returned metric.
-    """
-    from apps.import_log.models import ImportLog
-    from apps.media_library.models import MediaGallery, MediaGalleryItem, MediaItem
-
-    source = "viernulvier:media_item_gallery_links"
-    import_log = ImportLog.objects.create(
-        source=source,
-        status=ImportLog.Status.IN_PROGRESS,
-        started_at=timezone.now(),
+    return _media.sync_media_item_gallery_links_impl(
+        fetch_fn=fetch_viernulvier,
+        extract_external_id_fn=_extract_external_id_from_url,
+        timezone_module=timezone,
+        dry_run=dry_run,
+        etag_cache=etag_cache,
+        on_progress=on_progress,
+        params=params,
     )
-
-    try:
-        galleries = fetch_viernulvier(endpoint="/media/galleries", params=params, etag_cache=etag_cache)
-    except Exception as exc:
-        import_log.status = ImportLog.Status.FAILED
-        import_log.finished_at = timezone.now()
-        import_log.error_message = str(exc)
-        import_log.save()
-        raise
-
-    if not galleries:
-        import_log.status = ImportLog.Status.SUCCESS
-        import_log.records_total = 0
-        import_log.records_imported = 0
-        import_log.records_failed = 0
-        import_log.finished_at = timezone.now()
-        import_log.save()
-        return 0
-
-    gallery_pk_by_external_id = {str(ext_id): pk for ext_id, pk in MediaGallery.objects.values_list("external_id", "pk")}
-    media_pk_by_external_id = {str(ext_id): pk for ext_id, pk in MediaItem.objects.values_list("external_id", "pk")}
-
-    total = len(galleries)
-    errors = 0
-    error_messages: List[str] = []
-
-    gallery_item_links: List[MediaGalleryItem] = []
-    primary_item_to_gallery: Dict[int, int] = {}
-    primary_item_to_position: Dict[int, int] = {}
-    touched_gallery_ids: Set[int] = set()
-
-    def _record_error(msg: str) -> None:
-        nonlocal errors
-        errors += 1
-        if len(error_messages) < MAX_ERROR_MESSAGES:
-            error_messages.append(msg)
-
-    for idx, gallery_item in enumerate(galleries, start=1):
-        if not isinstance(gallery_item, dict):
-            _record_error(f"Gallery item is not a dict: {gallery_item!r}")
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        gallery_ext_id = _extract_external_id_from_url(gallery_item.get("@id"))
-        if not gallery_ext_id:
-            _record_error(f"Gallery missing @id: {str(gallery_item)[:200]}")
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        gallery_pk = gallery_pk_by_external_id.get(gallery_ext_id)
-        if gallery_pk is None:
-            _record_error(f"Gallery not found for external_id={gallery_ext_id}")
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        touched_gallery_ids.add(gallery_pk)
-        raw_items = gallery_item.get("items")
-        if not isinstance(raw_items, list):
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        for position, raw_media in enumerate(raw_items):
-            media_ext_id = _extract_external_id_from_url(raw_media)
-            if not media_ext_id:
-                continue
-            media_pk = media_pk_by_external_id.get(media_ext_id)
-            if media_pk is None:
-                _record_error(f"MediaItem not found for external_id={media_ext_id}")
-                continue
-
-            gallery_item_links.append(
-                MediaGalleryItem(
-                    gallery_id=gallery_pk,
-                    media_item_id=media_pk,
-                    position=position,
-                )
-            )
-
-            # Keep first occurrence as compatibility mirror for MediaItem.gallery.
-            if media_pk not in primary_item_to_gallery:
-                primary_item_to_gallery[media_pk] = gallery_pk
-                primary_item_to_position[media_pk] = position
-
-        if on_progress:
-            on_progress(idx, total)
-
-    linked_ids = set(primary_item_to_gallery.keys())
-
-    # Count MediaItem compatibility-field changes (clear + update).
-    media_items_to_clear = 0
-    media_items_to_update = 0
-    if touched_gallery_ids:
-        media_items_to_clear = (
-            MediaItem.objects.filter(gallery_id__in=touched_gallery_ids).exclude(pk__in=linked_ids).count()
-        )
-    if linked_ids:
-        for obj in MediaItem.objects.filter(pk__in=linked_ids).only("pk", "gallery_id", "position"):
-            new_gallery_id = primary_item_to_gallery[obj.pk]
-            new_position = primary_item_to_position[obj.pk]
-            if obj.gallery_id != new_gallery_id or obj.position != new_position:
-                media_items_to_update += 1
-
-    media_items_changed = media_items_to_clear + media_items_to_update
-    link_rows_written = len(gallery_item_links)
-
-    if dry_run:
-        import_log.status = ImportLog.Status.SUCCESS if errors == 0 else ImportLog.Status.PARTIAL_SUCCESS
-        import_log.records_total = total
-        import_log.records_imported = media_items_changed
-        import_log.records_failed = errors
-        import_log.finished_at = timezone.now()
-        if errors:
-            import_log.error_message = f"{errors} link issues: {', '.join(error_messages)}"
-        import_log.save()
-        logger.info(
-            "Gallery-item link sync complete: media_items_changed=%d link_rows=%d errors=%d [DRY RUN]",
-            media_items_changed,
-            link_rows_written,
-            errors,
-        )
-        return media_items_changed
-
-    actual_link_rows_written = 0
-    actual_media_items_changed = 0
-
-    exc: Optional[BaseException] = None
-    try:
-        with transaction.atomic():
-            if touched_gallery_ids:
-                # Replace links for galleries that were part of this sync run.
-                MediaGalleryItem.objects.filter(gallery_id__in=touched_gallery_ids).delete()
-                if gallery_item_links:
-                    MediaGalleryItem.objects.bulk_create(gallery_item_links)
-                actual_link_rows_written = len(gallery_item_links)
-
-                # Keep MediaItem.gallery in sync as a best-effort compatibility field.
-                cleared = (
-                    MediaItem.objects.filter(gallery_id__in=touched_gallery_ids)
-                    .exclude(pk__in=linked_ids)
-                    .update(gallery_id=None)
-                )
-                actual_media_items_changed += cleared
-
-            to_update = []
-            for obj in MediaItem.objects.filter(pk__in=linked_ids).only("pk", "gallery_id", "position"):
-                new_gallery_id = primary_item_to_gallery[obj.pk]
-                new_position = primary_item_to_position[obj.pk]
-                if obj.gallery_id != new_gallery_id or obj.position != new_position:
-                    obj.gallery_id = new_gallery_id
-                    obj.position = new_position
-                    to_update.append(obj)
-
-            if to_update:
-                MediaItem.objects.bulk_update(to_update, ["gallery", "position"])
-                actual_media_items_changed += len(to_update)
-    except Exception as e:
-        logger.exception("Error while syncing media gallery links")
-        exc = e
-    finally:
-        if exc is not None:
-            import_log.status = ImportLog.Status.FAILED
-            import_log.records_total = total
-            import_log.records_imported = actual_media_items_changed
-            import_log.records_failed = errors + 1
-            import_log.finished_at = timezone.now()
-            import_log.error_message = f"Exception during gallery link sync: {exc}"
-            import_log.save()
-    if exc is not None:
-        raise exc
-
-    import_log.records_total = total
-    import_log.records_imported = actual_media_items_changed
-    import_log.records_failed = errors
-    import_log.finished_at = timezone.now()
-
-    if errors == 0:
-        import_log.status = ImportLog.Status.SUCCESS
-    elif actual_media_items_changed > 0:
-        import_log.status = ImportLog.Status.PARTIAL_SUCCESS
-        import_log.error_message = f"{errors} link issues: {', '.join(error_messages)}"
-    else:
-        import_log.status = ImportLog.Status.FAILED
-        import_log.error_message = f"All {errors} link resolutions failed: {', '.join(error_messages)}"
-
-    import_log.save()
-    logger.info(
-        "Gallery-item link sync complete: media_items_changed=%d link_rows=%d errors=%d",
-        actual_media_items_changed,
-        actual_link_rows_written,
-        errors,
-    )
-    return actual_media_items_changed
-
-
-# ---------------------------------------------------------------------------
-# Media item crop sync
-# ---------------------------------------------------------------------------
-
-
-def _derive_crop_filename(crop_name: str, item_external_id: str, image_url: str) -> str:
-    """Build a deterministic local filename for a crop image.
-
-    Pattern: ``<item_id_slug>_<crop_name>.<ext>``
-    Example: ``api_v1_media_items_310_hd_ready.jpg``
-    """
-    slug = item_external_id.strip("/").replace("/", "_")
-    # Extract extension from the CDN URL path (before any query string)
-    path_part = image_url.split("?")[0]
-    last_segment = path_part.split("/")[-1]
-    if "." in last_segment:
-        ext = "." + last_segment.rsplit(".", 1)[-1].lower()
-        # CDN URLs often end in a hash with no real extension; fall back to .jpg
-        if len(ext) > 5 or not ext[1:].isalpha():
-            ext = ".jpg"
-    else:
-        ext = ".jpg"
-    return f"{slug}_{crop_name}{ext}"
-
-
-def _download_image(session: requests.Session, url: str) -> Optional[bytes]:
-    """Download image bytes from a CDN URL with retry + backoff.
-
-    Returns None on permanent failure so the caller can skip this crop
-    without aborting the whole sync.
-    """
-    for attempt in range(MAX_RETRIES + 1):
-        try:
-            response = session.get(url, timeout=30, stream=True)
-        except (requests.ConnectionError, requests.Timeout) as exc:
-            if attempt == MAX_RETRIES:
-                logger.error("Image download failed after %d retries: %s (%s)", MAX_RETRIES, url, exc)
-                return None
-            wait = _backoff_seconds(attempt)
-            logger.warning("Image download error - retry %d in %.1fs: %s", attempt + 1, wait, url)
-            time.sleep(wait)
-            continue
-        except requests.RequestException as exc:
-            logger.error("Image download request error: %s (%s)", url, exc)
-            return None
-
-        if response.status_code == 429:
-            wait = float(response.headers.get("Retry-After", _backoff_seconds(attempt)))
-            if attempt == MAX_RETRIES:
-                logger.error("Rate limited downloading image (gave up): %s", url)
-                return None
-            logger.warning("Rate limited downloading image - waiting %.1fs: %s", wait, url)
-            time.sleep(wait)
-            continue
-
-        if not response.ok:
-            logger.error("HTTP %d downloading image: %s", response.status_code, url)
-            return None
-
-        return response.content
-
-    return None
 
 
 def sync_media_item_crops(
@@ -1520,331 +231,16 @@ def sync_media_item_crops(
     on_progress: Optional[Callable[[int, int], None]] = None,
     params: Optional[Dict[str, str]] = None,
 ) -> int:
-    """Fetch individual foto MediaItems to obtain crop data, download images.
-
-    Strategy:
-    1. Build the list of candidate media items.
-       - With created/updated filters: fetch `/media/items` from the API and use
-         those filtered results directly (independent of local DB state).
-       - Without filters: fall back to local foto items for performance.
-    2. For each candidate, fetch its full API representation (includes crops).
-    3. Ensure a local MediaItem exists before writing crops.
-       Crops depend on MediaItem via FK, so missing items are upserted first.
-    4. For each wanted crop variant (hd_ready, FE3_header):
-       a. Download the image bytes.
-       b. Upsert the MediaItemCrop row, saving the image via Django's storage
-          backend. The file is written to storage before the DB row is touched
-          so the ImageField is never blank on INSERT.
-
-    Only foto items are processed because the API only embeds crops for
-    image-type assets, and fetching every item individually is expensive.
-
-    Args:
-        dry_run:     When True, log what would be saved but write nothing.
-        on_progress: Callback(processed_items, total_items) invoked after each
-                     media item is handled (regardless of crop count).
-        params:      Optional API query parameters for timestamp filtering
-                     (e.g. created_at[after], updated_at[after]).
-
-    Returns:
-        Number of crops saved (created + updated). Always 0 in dry_run mode.
-    """
-    from apps.import_log.models import ImportLog
-    from apps.media_library.models import MediaGallery, MediaItem, MediaItemCrop
-
-    source = "viernulvier:media_item_crops"
-    import_log = ImportLog.objects.create(
-        source=source,
-        status=ImportLog.Status.IN_PROGRESS,
-        started_at=timezone.now(),
+    return _media.sync_media_item_crops_impl(
+        fetch_fn=fetch_viernulvier,
+        build_session_fn=_build_session,
+        fetch_with_retry_fn=_fetch_with_retry,
+        download_image_fn=_download_image,
+        derive_crop_filename_fn=_derive_crop_filename,
+        extract_external_id_fn=_extract_external_id_from_url,
+        parse_datetime_fn=parse_datetime,
+        timezone_module=timezone,
+        dry_run=dry_run,
+        on_progress=on_progress,
+        params=params,
     )
-
-    # Start from local foto items (fast path), then replace with API-driven
-    # candidates when timestamp filters are provided.
-    foto_items_query = MediaItem.objects.filter(type=MediaItem.MediaItemType.IMAGE)
-    api_filter_params: Dict[str, str] = {}
-
-    if params:
-        # Parse timestamp params. API filtering is independent from local model
-        # fields; local filtering remains best-effort as a pre-filter only.
-        for param_key, param_value in params.items():
-            match = re.fullmatch(r"(created_at|updated_at)\[(after|before|strictly_after|strictly_before)\]", param_key)
-            if not match:
-                continue
-
-            field_name, bound = match.groups()
-            api_filter_params[param_key] = param_value
-
-            try:
-                MediaItem._meta.get_field(field_name)
-            except FieldDoesNotExist:
-                logger.debug(
-                    "Skipping local pre-filter '%s' for media_item_crops (field '%s' not on MediaItem)",
-                    param_key,
-                    field_name,
-                )
-                continue
-
-            try:
-                dt = parse_datetime(param_value)
-            except (ValueError, TypeError):
-                continue
-
-            if not dt:
-                continue
-
-            lookup = "gte" if bound in {"after", "strictly_after"} else "lte"
-            foto_items_query = foto_items_query.filter(**{f"{field_name}__{lookup}": dt})
-
-    foto_items = list(foto_items_query.values("pk", "external_id"))
-
-    if api_filter_params:
-        # API-driven filtered runs: do not depend on local MediaItem existence.
-        try:
-            api_items = fetch_viernulvier(endpoint="/media/items", params=api_filter_params)
-        except Exception as exc:
-            import_log.status = ImportLog.Status.FAILED
-            import_log.finished_at = timezone.now()
-            import_log.error_message = str(exc)
-            import_log.save()
-            raise
-
-        existing_pk_by_external_id = {
-            str(row["external_id"]).strip(): row["pk"] for row in foto_items if row.get("external_id")
-        }
-        api_candidates: List[Dict[str, Any]] = []
-        for api_item in api_items:
-            if not isinstance(api_item, dict):
-                continue
-
-            raw_type = str(api_item.get("type") or "").strip().lower()
-            if raw_type and raw_type != MediaItem.MediaItemType.IMAGE:
-                continue
-
-            ext_id = _extract_external_id_from_url(api_item.get("@id") or api_item.get("external_id") or api_item.get("id"))
-            if ext_id:
-                normalized_ext_id = str(ext_id).strip()
-                api_candidates.append(
-                    {
-                        "pk": existing_pk_by_external_id.get(normalized_ext_id),
-                        "external_id": normalized_ext_id,
-                    }
-                )
-
-        before_count = len(foto_items)
-        foto_items = api_candidates
-
-        logger.info(
-            "Applied API filters to media_item_crops candidates: %d -> %d (local-independent)",
-            before_count,
-            len(foto_items),
-        )
-
-    if not foto_items:
-        import_log.status = ImportLog.Status.SUCCESS
-        import_log.records_total = 0
-        import_log.records_imported = 0
-        import_log.records_failed = 0
-        import_log.finished_at = timezone.now()
-        import_log.save()
-        logger.info("No foto MediaItems found - skipping crop sync.")
-        return 0
-
-    wanted_crops: Set[str] = MediaItemCrop.SYNCED_CROP_NAMES
-    total = len(foto_items)
-    saved = 0
-    errors = 0
-    error_messages: List[str] = []
-    session = _build_session()
-
-    logger.info(
-        "Syncing crops for %d foto MediaItems (variants: %s)",
-        total,
-        ", ".join(sorted(wanted_crops)),
-    )
-
-    for idx, row in enumerate(foto_items, start=1):
-        item_pk = row.get("pk")
-        external_id: str = row["external_id"] or ""
-
-        if not external_id:
-            logger.warning("MediaItem pk=%d has no external_id - skipping", item_pk)
-            errors += 1
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        # external_id is already a path like "/api/v1/media/items/310"
-        item_url = external_id if external_id.startswith("http") else urljoin(BASE_DOMAIN, external_id)
-
-        try:
-            item_data, _ = _fetch_with_retry(session, item_url)
-        except ScraperError as exc:
-            logger.error("Failed to fetch media item %s: %s", item_url, exc)
-            errors += 1
-            if len(error_messages) < MAX_ERROR_MESSAGES:
-                error_messages.append(f"Fetch failed for {external_id}: {exc}")
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        if item_data is None:
-            # 304 Not Modified - crops unchanged, nothing to do
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        if item_pk is None:
-            # Crops depend on MediaItem via FK. If the row is missing locally,
-            # create/update it from this same payload before persisting crops.
-            if dry_run:
-                logger.info("[DRY RUN] Would upsert missing MediaItem for crop sync: %s", external_id)
-            else:
-                try:
-                    media_type_raw = str(item_data.get("type") or MediaItem.MediaItemType.IMAGE).strip().lower()
-                    valid_types = {choice for choice, _ in MediaItem.MediaItemType.choices}
-                    media_type = media_type_raw if media_type_raw in valid_types else MediaItem.MediaItemType.IMAGE
-
-                    gallery_ext_id = _extract_external_id_from_url(item_data.get("gallery"))
-                    gallery_pk = None
-                    if gallery_ext_id:
-                        gallery_pk = (
-                            MediaGallery.objects.filter(external_id=str(gallery_ext_id).strip())
-                            .values_list("pk", flat=True)
-                            .first()
-                        )
-
-                    obj, _ = MediaItem.objects.update_or_create(
-                        external_id=external_id,
-                        defaults={
-                            "type": media_type,
-                            "gallery_id": gallery_pk,
-                            "original_filename": str(item_data.get("original_filename") or ""),
-                            "position": item_data.get("position") or 0,
-                            "width": item_data.get("width"),
-                            "height": item_data.get("height"),
-                            "format": str(item_data.get("format") or ""),
-                        },
-                    )
-                    item_pk = obj.pk
-                    logger.info(
-                        "Upserted missing MediaItem dependency for crops: %s (pk=%s)",
-                        external_id,
-                        item_pk,
-                    )
-                except Exception as exc:
-                    logger.exception("Failed to upsert MediaItem dependency for crop sync (%s): %s", external_id, exc)
-                    errors += 1
-                    if len(error_messages) < MAX_ERROR_MESSAGES:
-                        error_messages.append(f"Missing MediaItem upsert failed for {external_id}: {exc}")
-                    if on_progress:
-                        on_progress(idx, total)
-                    continue
-
-        if item_pk is None:
-            logger.error("MediaItem dependency unresolved for crop sync: %s", external_id)
-            errors += 1
-            if len(error_messages) < MAX_ERROR_MESSAGES:
-                error_messages.append(f"MediaItem dependency unresolved for {external_id}")
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        crops_raw = item_data.get("crops", [])
-        if not isinstance(crops_raw, list):
-            logger.debug("Unexpected crops format for %s: %r", external_id, crops_raw)
-            if on_progress:
-                on_progress(idx, total)
-            continue
-
-        for crop_data in crops_raw:
-            if not isinstance(crop_data, dict):
-                continue
-
-            crop_name: str = crop_data.get("name", "")
-            if crop_name not in wanted_crops:
-                continue
-
-            image_url: str = crop_data.get("url", "")
-            if not image_url:
-                logger.warning("Crop '%s' for %s has no URL - skipping", crop_name, external_id)
-                continue
-
-            if dry_run:
-                logger.info(
-                    "[DRY RUN] Would save crop '%s' for MediaItem pk=%d from %s",
-                    crop_name,
-                    item_pk,
-                    image_url,
-                )
-                saved += 1
-                continue
-
-            image_bytes = _download_image(session, image_url)
-            if image_bytes is None:
-                errors += 1
-                if len(error_messages) < MAX_ERROR_MESSAGES:
-                    error_messages.append(f"Download failed for crop '{crop_name}' on {external_id}")
-                continue
-
-            filename = _derive_crop_filename(crop_name, external_id, image_url)
-
-            try:
-                # Write the file to storage *before* touching the DB so the
-                # ImageField is never blank on INSERT (blank=False constraint).
-                image_field = MediaItemCrop._meta.get_field("image")
-                upload_name = image_field.generate_filename(None, filename)
-                saved_path = image_field.storage.save(upload_name, ContentFile(image_bytes))
-
-                with transaction.atomic():
-                    _, created = MediaItemCrop.objects.update_or_create(
-                        media_item_id=item_pk,
-                        name=crop_name,
-                        defaults={"image": saved_path},
-                    )
-
-                saved += 1
-                logger.debug(
-                    "%s crop '%s' for MediaItem pk=%d -> %s",
-                    "Created" if created else "Updated",
-                    crop_name,
-                    item_pk,
-                    saved_path,
-                )
-            except Exception as exc:
-                logger.exception(
-                    "Error saving crop '%s' for MediaItem pk=%d: %s",
-                    crop_name,
-                    item_pk,
-                    exc,
-                )
-                errors += 1
-                if len(error_messages) < MAX_ERROR_MESSAGES:
-                    error_messages.append(f"Save failed for crop '{crop_name}' on {external_id}: {exc}")
-
-        if on_progress:
-            on_progress(idx, total)
-
-    # Finalise import log
-    import_log.records_total = total
-    import_log.records_imported = saved
-    import_log.records_failed = errors
-    import_log.finished_at = timezone.now()
-
-    if errors == 0:
-        import_log.status = ImportLog.Status.SUCCESS
-    elif saved > 0:
-        import_log.status = ImportLog.Status.PARTIAL_SUCCESS
-        import_log.error_message = f"{errors} crops failed: {', '.join(error_messages)}"
-    else:
-        import_log.status = ImportLog.Status.FAILED
-        import_log.error_message = f"All {errors} crops failed: {', '.join(error_messages)}"
-
-    import_log.save()
-    logger.info(
-        "Crop sync complete: saved=%d, errors=%d%s",
-        saved,
-        errors,
-        " [DRY RUN]" if dry_run else "",
-    )
-    return saved
